+++ Findstaff/ucJobCategory.cs
@@ -52,11 +52,15 @@ namespace Findstaff
             Connection con = new Connection();
             connection = con.dbConnection();
             connection.Open();
-            string cmd = "delete from jobcategory_t where category_id = '" + dgvJobCategory.SelectedRows[0].Cells[0].Value.ToString() + "';";
-            com = new MySqlCommand(cmd, connection);
-            com.ExecuteNonQuery();
-            dgvJobCategory.Rows.Remove(dgvJobCategory.SelectedRows[0]);
-            MessageBox.Show("Job Category Deleted!", "Job Category Record Removed", MessageBoxButtons.OK, MessageBoxIcon.Information);
+            DialogResult r = MessageBox.Show("Do you want to delete the category " + dgvJobCategory.SelectedRows[0].Cells[1].Value.ToString() + "?", "Delete Skill confirmation", MessageBoxButtons.YesNo, MessageBoxIcon.Question);
+            if (r == DialogResult.Yes)
+            {
+                string cmd = "delete from jobcategory_t where category_id = '" + dgvJobCategory.SelectedRows[0].Cells[0].Value.ToString() + "';";
+                com = new MySqlCommand(cmd, connection);
+                com.ExecuteNonQuery();
+                dgvJobCategory.Rows.Remove(dgvJobCategory.SelectedRows[0]);
+                MessageBox.Show("Job Category Deleted!", "Job Category Record Removed", MessageBoxButtons.OK, MessageBoxIcon.Information);
+            }
             connection.Close();
         }
